fix: Use correct .zshrc location when $ZDOTDIR set
Spawn a new shell process to get the value of $ZDOTDIR, if set. This is needed so Homebrew will install correctly for users who have moved their zsh config directory. Note that we cannot simply read this from the environment, as one of the common methods of setting this value (via ~/.zshenv), results in it being a shell variable and not entering the environment.